Add proper loading state with changed routing + other changes
