+++ app/boards/[boardId]/page.tsx
@@ -1,19 +1,18 @@
 'use client'
+import { LoadingSpinner } from '@/components/global/LoadingSpinner'
 import { Board } from '@/components/planner/Board/Board'
 import { usePlanner } from '@/hooks/Planner/Planner'
 
-interface BoardPageProps {
-  params: {
-    boardId: string
-  }
-}
-
-export default function BoardPage({ params }: BoardPageProps) {
+export default function BoardPage({ params }: { params: { boardId: string } }) {
   const { hasLoaded } = usePlanner()
   const { boardId } = params
 
   if (!hasLoaded) {
-    return <></>
+    return (
+      <div className='flex flex-col justify-center items-center gap-2 w-full'>
+        <LoadingSpinner />
+      </div>
+    )
   }
   return <Board boardId={boardId} />
 }
