--- components/global/LoadingSpinner/LoadingSpinner.module.css
@@ -1,14 +1,0 @@
-@keyframes loader {
-  100% {
-    transform: rotate(360deg);
-  }
-}
-
-.loadingSpinner {
-  width: 50px;
-  height: 50px;
-  border: 10px solid #eee;
-  border-top: 10px solid #363b3e;
-  border-radius: 100%;
-  animation: loader 1s infinite linear;
-}
